% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_np.R
\name{as.numeric_np}
\alias{as.numeric_np}
\title{Apply numeric_np class to a numeric variable}
\usage{
as.numeric_np(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
To allow \code{atable} to return different statistics for different variables
of the same type (e.g. we want some variables to be formatted as
\code{Mean (SD)} and others as \code{Median [IQR]}), they need to be passed
to \code{atable} as different classes.
}
\examples{
data(mtcars)
mtcars$wt_np <- as.numeric_np(mtcars$wt)

}
