% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_statistics.R
\name{format_statistics.numeric_stats}
\alias{format_statistics.numeric_stats}
\title{Format summary statistics for numeric variables when using `numeric_stats`}
\usage{
\method{format_statistics}{numeric_stats}(
  x,
  numstats = c(`Mean (SD)` = "{mean} ({sd})", `Min - Max` = "{q0} - {q1}", `Median
    [Quartiles]` = "{q0.5} [{q0.25}; {q0.75}]")[1],
  missingformat = c(`Valid (missing)` = "{Nvalid} ({Nmissing})"),
  ...
)
}
\arguments{
\item{x}{`numeric_stats` object}

\item{numstats}{either a number (1 to 3, referring to the defaults in the order printed) or a named character of the format shown above (i.e. 'Text to print in table' = '{variable1 to use} ({variable2 to use})'). Defaults to `"Mean (SD)" = "{mean} ({sd})"`, equivilent to 1.}

\item{missingformat}{either a logical or a named string defining the format}
}
\value{

}
\description{
format_statistics.numeric_stats can use any statistic returned by `numeric_stats`.
}
\details{

}
\examples{

require(atable)
atable_options(statistics.numeric = numeric_stats)
data(mtcars)
# atableExtra default
atable(mpg ~ am, mtcars)
# specify only mean
atable(mpg ~ am, mtcars, numstats = c("Mean (SD)" = "{mean} ({sd})"))
# specify options by number
atable(mpg ~ am, mtcars, numstats = 1:2)
atable(mpg ~ am, mtcars, numstats = c(1,3))
# change to square brackets
atable(mpg ~ am, mtcars, numstats = c("Mean [SD]" = "{mean} [{sd}]"))

}
