% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_stats.R
\name{numeric_stats}
\alias{numeric_stats}
\title{Calculate summary statistics for numeric variables}
\usage{
numeric_stats(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{passed to/from other methods}

\item{quantiles}{list of quantiles to calculate}
}
\value{
list of class `numeric_stats` with one element for each quantile
(e.g. `q0.5` for median), iqr (interquartile range, only if quantiles includes
0.25 and 0.75), range (only if quartiles includes 0 and 1),
mean, sd, N (length), Nvalid (number of non-`NA`s) and Nmissing (number of `NA`s)
}
\description{
This function is designed to be compatible with `atable` as a substitute for
it's default function (`atable:::statistics.numeric()`). It returns mean,
SD, any specified quantiles, number of observations, number of valid and missing
values. Together with the appropriate formatting function, This allows a simple
and flexible approach to defining the formats.
}
\details{

}
\examples{
require(atable)
data(mtcars)
# single use (specify in atable call)
atable(mpg ~ am, mtcars, statistics.numeric = numeric_stats)
# or change atables default
atable_options(statistics.numeric = numeric_stats)
atable(mpg ~ am, mtcars)
}
\seealso{
format_statistics.numeric_stats
}
