% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_statistics.R
\name{format_statistics.numeric_stats}
\alias{format_statistics.numeric_stats}
\alias{format_statistics.numeric_stats_np}
\title{Format summary statistics for numeric variables when using \code{numstats}}
\usage{
\method{format_statistics}{numeric_stats}(
  x,
  numstats = c(`Mean (SD)` = "{mean} ({sd})"),
  missingformat = c(`Valid (missing)` = "{Nvalid} ({Nmissing})"),
  ...
)

\method{format_statistics}{numeric_stats_np}(
  x,
  np_numstats = 2,
  np_missingformat = c(`Valid (missing)` = "{Nvalid} ({Nmissing})"),
  ...
)
}
\arguments{
\item{x}{a \code{numstats} object}

\item{numstats}{either a number (1 to 3) or a named character, where the name
is used as the row name and the second part in the formatting. See details for
more info.}

\item{missingformat}{either a logical or a named string defining the format
as for \code{numstats}.}

\item{...}{passed to/from other methods}
}
\description{
format_statistics.numeric_stats can use any statistic returned by \code{numstats}.
It returns the type of table expected by \code{atable}.
}
\details{
Options \code{numstats} and \code{missingformat} use the \code{glue} package
to create strings. Although \code{glue} parses the values within the braces
(`\{` and `\}`), this is not permitted in this case as only strings are passed
to \code{glue} after being formatted by \code{atable_options("format_numbers")}.
To format the numbers differently, reassign that function.

The \code{numstats} argument allows the flexible formatting of the statistics
and label. It accepts either a number (1 to 3) or a names character string.
Number 1 is equivalent to \code{c("Mean (SD)" = "{mean} ({sd})")}. Number 2
is equivalent to \code{c("Median [Quartiles]" = "{q0.5} [{q0.25}; {q0.75}]")}
and number 3 is equivalent to \code{c("Min - Max" = "{q0} - {q1}")}.
Any parameter returned by \code{numeric_stats} can be used.

\code{missingformat} works in the same way as \code{numstats}. Tt defaults to
\code{c("Valid (missing)" = "{Nvalid} ({Nmissing})")}.
}
\section{Functions}{
\itemize{
\item \code{format_statistics.numeric_stats_np}: Method for numeric_stats_np class
}}

\examples{

require(atable)
atable_options(statistics.numeric = numeric_stats)
data(mtcars)
# atableExtra default
atable(mpg ~ am, mtcars)
# specify only mean
atable(mpg ~ am, mtcars, numstats = c("Mean (SD)" = "{mean} ({sd})"))
# specify options by number
atable(mpg ~ am, mtcars, numstats = 1:2)
atable(mpg ~ am, mtcars, numstats = c(1,3))
# change to square brackets
atable(mpg ~ am, mtcars, numstats = c("Mean [SD]" = "{mean} [{sd}]"))

}
