% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_statistics.R
\name{format_statistics.numeric_stats}
\alias{format_statistics.numeric_stats}
\title{Format summary statistics for numeric variables when using `numeric_stats`}
\usage{
\method{format_statistics}{numeric_stats}(
  x,
  numstats = c(`Mean (SD)` = "{mean} ({sd})"),
  missingformat = c(`Valid (missing)` = "{Nvalid} ({Nmissing})"),
  ...
)
}
\arguments{
\item{x}{`numeric_stats` object}

\item{numstats}{either a number (1 to 3, referring to the defaults in the order printed) or a named character of the format shown above (i.e. 'Text to print in table' = '{variable1 to use} ({variable2 to use})'), where the name is used as the row name and the second part in the formatting. Defaults to `"Mean (SD)" = "{mean} ({sd})"`, equivilent to 1. `numstats` = 2 is equivalent to \code("Min - Max" = "{q0} - {q1}") and 3 is equivalent to \code("Median [Quartiles]" = "{q0.5} [{q0.25}; {q0.75}]")}

\item{missingformat}{either a logical or a named string defining the format as for `numstats`.}

\item{...}{passed to/from other methods}
}
\description{
format_statistics.numeric_stats can use any statistic returned by `numeric_stats`.
}
\details{
All numbers are formatted by the function in `atable_options("format_numbers")`.
To use a different method, reassign that function.

Options `numstats` and `missingformat` use the `glue` package to create strings.
Although `glue` parses the values within the braces (`{` and `}`), this is not permitted
in this case as only strings are passed to `glue`.
}
\examples{

require(atable)
atable_options(statistics.numeric = numeric_stats)
data(mtcars)
# atableExtra default
atable(mpg ~ am, mtcars)
# specify only mean
atable(mpg ~ am, mtcars, numstats = c("Mean (SD)" = "{mean} ({sd})"))
# specify options by number
atable(mpg ~ am, mtcars, numstats = 1:2)
atable(mpg ~ am, mtcars, numstats = c(1,3))
# change to square brackets
atable(mpg ~ am, mtcars, numstats = c("Mean [SD]" = "{mean} [{sd}]"))

}
