% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_statistics.R
\name{format_statistics.statistics_factor}
\alias{format_statistics.statistics_factor}
\title{Format factors as `X (X%)` instead of `X% (X)`}
\usage{
\method{format_statistics}{statistics_factor}(x, factorformat = "{N} ({percent}\%)", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Almost an exact copy of the equivilent function from `atable`,
this function just formats factors as e.g. 15 (25%) rather than
25% (15) as `atable` does.
}
