% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_stats.R
\name{numeric_stats}
\alias{numeric_stats}
\title{Calculate summary statistics for numeric variables
This function is designed to be compatible with `atable` as a substitute for
it's default function (`atable:::statistics.numeric()`).}
\usage{
numeric_stats(x, quantiles = c(0, 0.25, 0.5, 0.75, 1), ...)
}
\arguments{
\item{x}{data}

\item{quantiles}{list of quantiles to calculate}
}
\value{
list of class `numeric_stats` with one element for each quantile (e.g. `q0.5` for median),
mean, sd, N (length), Nvalid (number of non-`NA`s) and Nmissing (number of `NA`s)
}
\description{
Calculate summary statistics for numeric variables
This function is designed to be compatible with `atable` as a substitute for
it's default function (`atable:::statistics.numeric()`).
}
\details{

}
\examples{
require(atable)
data(mtcars)
# single use (specify in atable call)
atable(mpg ~ am, mtcars, statistics.numeric = numeric_stats)
# or change atables default
atable_options(statistics.numeric = numeric_stats)
atable(mpg ~ am, mtcars)

}
